feat(core): implement search logic

+++ mintaka_trainer/arena/arena.py
@@ -19,6 +19,7 @@ class Color(Enum):
     def __str__(self) -> str:
         return self.name.lower()
 
+
 class Player(Enum):
     A = 0
     B = 1
@@ -29,9 +30,17 @@ class Player(Enum):
     def __str__(self) -> str:
         return self.name
 
+
 class Config:
     def __init__(self, args):
         self.args = args
+
+
+@dataclass
+class Opening:
+    initial_color: Color
+    sequence: str
+
 
 @dataclass
 class TimeManager:
@@ -44,6 +53,7 @@ class TimeManager:
 
     def consume(self, running_time):
         self.total_remaining = max(0, self.total_remaining - running_time)
+
 
 @dataclass
 class GameResult:
@@ -63,15 +73,18 @@ class GameResult:
     def win_zero_to_one(self, player) -> float:
         return self._win(self.color[player])
 
+
 @dataclass
 class Engine:
     process: subprocess.Popen
     time_manager: TimeManager
 
+
 def calculate_elo_delta(original_elo, opponent_elo, result_zero_to_one, k_factor) -> float:
     expected = 1.0 / (1.0 + 10.0 ** ((opponent_elo - original_elo) / 400.0))
 
     return k_factor * (result_zero_to_one - expected)
+
 
 def datetime_prefix(config) -> str:
     if config.args.no_datetime_prefix:
@@ -79,24 +92,28 @@ def datetime_prefix(config) -> str:
     else:
         return datetime.now(timezone.utc).strftime("[%Y-%m-%dT%H:%M:%SZ] ")
 
+
 def game_prefix(config, game_no) -> str:
     return f"[{game_no + 1}/{config.args.num_games}] "
+
 
 def turn_prefix(turn_no, player, color) -> str:
     return f"[#{turn_no}:{player}:{color}] "
 
-def spawn_process(path, params, time_manager, opening) -> subprocess.Popen:
+
+def spawn_process(path, params, time_manager: TimeManager, opening: Opening | None) -> subprocess.Popen:
     return subprocess.Popen(
         [path]
             + shlex.split(params)
             + ["--time", str(time_manager.total_remaining), str(time_manager.increment), str(time_manager.turn)]
-            + (["--opening", opening] if opening is not None else []),
+            + (["--history", opening.sequence] if opening is not None else []),
         stdin=subprocess.PIPE,
         stdout=subprocess.PIPE,
         stderr=subprocess.STDOUT,
         text=True,
         bufsize=1
     )
+
 
 def command_process(config, process, command,
                     println=False, filter_prefix: list[str] | None = None, println_prefix: str = "") -> str | None:
@@ -130,12 +147,15 @@ def command_process(config, process, command,
         if response.startswith("="):
             return response[2:]
 
-        if println and (filter_prefix is None or response.startswith(tuple(filter_prefix))):
+        if not config.args.concise and println and (filter_prefix is None or response.startswith(tuple(filter_prefix))):
             print(f"{datetime_prefix(config)}{println_prefix}{response}")
 
-def play_game(config, game_no, opening) -> GameResult:
+
+def play_game(config, game_no: int, opening: Opening | None) -> GameResult:
     a_time_manager = TimeManager(config.args.a_time[0], config.args.a_time[1], config.args.a_time[2])
     b_time_manager = TimeManager(config.args.b_time[0], config.args.b_time[1], config.args.b_time[2])
+
+    initial_color = opening.initial_color if opening is not None else Color.BLACK
 
     with (
         spawn_process(config.args.a_path, config.args.a_params, a_time_manager, opening) as a_process,
@@ -149,8 +169,8 @@ def play_game(config, game_no, opening) -> GameResult:
         player = Player.A if game_no % 2 == 0 else Player.B
 
         color = {
-            player: Color.BLACK,
-            player.flip(): Color.WHITE
+            player: initial_color,
+            player.flip(): initial_color.flip()
         }
 
         winner = None
@@ -173,13 +193,14 @@ def play_game(config, game_no, opening) -> GameResult:
             engines[player].time_manager.consume(time_elapsed)
 
             if move == "none":
-                winner_player = None
-            else:
-                winner_player = command_process(config, engines[player].process, f"play {move}")
-                winner_opponent = command_process(config, engines[player.flip()].process, f"play {move}")
+                winner = color[player.flip()]
+                break
 
-                if winner_player != winner_opponent:
-                    raise Exception(f"player_winner={winner_player}, opponent_winner={winner_opponent} ")
+            winner_player = command_process(config, engines[player].process, f"play {move}")
+            winner_opponent = command_process(config, engines[player.flip()].process, f"play {move}")
+
+            if winner_player != winner_opponent:
+                raise Exception(f"player_winner={winner_player}, opponent_winner={winner_opponent} ")
 
             if winner_player is not None:
                 if "black" in winner_player.lower():
@@ -199,12 +220,16 @@ def play_game(config, game_no, opening) -> GameResult:
             player = player.flip()
 
         history = command_process(config, engines[Player.A].process, "history")
-        board_str = command_process(config, engines[Player.B].process, "board", println=True)
+        board_str = command_process(config, engines[Player.A].process, "board", println=True)
+
+        if board_str is None or history is None:
+            raise Exception("board or history response is None")
 
         for engine in engines.values():
             engine.process.terminate()
 
         return GameResult(color, winner, history, board_str)
+
 
 def main():
     random.seed(42)
@@ -226,7 +251,7 @@ def main():
     parser.add_argument("--b-time", type=int, nargs=3, default=default_total_increment_turn,
                         help="total(ms) increment(ms) turn(ms)")
 
-    parser.add_argument("--openings", type=str, default=None, help="opening.csv, history on each line")
+    parser.add_argument("--openings-file", type=str, default=None, help="openings.csv, history on each line")
 
     parser.add_argument("--draw-in", type=int, default=225)
 
@@ -234,15 +259,22 @@ def main():
     parser.add_argument("--b-elo", type=float, default=1000.0)
     parser.add_argument("--elo-k-factor", type=float, default=32.0)
 
-    parser.add_argument("--no-datetime-prefix", action='store_true', default=False)
+    parser.add_argument("--concise", action="store_true", default=False)
+    parser.add_argument("--no-datetime-prefix", action="store_true", default=False)
     parser.add_argument("--log-prefix-filter", type=str, nargs="*", default=["solution"])
 
     config = Config(parser.parse_args())
 
     openings = []
-    if config.args.openings is not None:
-        with open(config.args.openings, "r") as openings_file:
-            openings = openings_file.read().splitlines()
+    if config.args.openings_file is not None:
+        with open(config.args.openings_file, "r") as openings_file:
+            openings = [
+                Opening(Color.BLACK if len(opening.split(",")) % 2 == 0 else Color.WHITE, opening)
+                for opening in openings_file.read().strip().splitlines()
+            ]
+
+            random.shuffle(openings)
+            openings = [opening for opening in openings for _ in range(2)]
 
     elo = {
         Player.A: config.args.a_elo,
@@ -259,7 +291,7 @@ def main():
     draws = 0
 
     for game_no in range(config.args.num_games):
-        result = play_game(config, game_no, random.choice(openings) if len(openings) != 0 else None)
+        result = play_game(config, game_no, openings[game_no % len(openings)] if len(openings) != 0 else None)
 
         elo_delta = calculate_elo_delta(
             elo[Player.A], elo[Player.B],
@@ -272,15 +304,17 @@ def main():
         wins[Player.A] += int(round(result.win_zero_to_one(Player.A)))
         wins[Player.B] += int(round(result.win_zero_to_one(Player.B)))
 
+        draws = game_no + 1 - wins[Player.A] - wins[Player.B]
+
         wins[Color.BLACK] += 1 if result.winner == Color.BLACK else 0
         wins[Color.WHITE] = game_no + 1 - wins[Color.BLACK] - draws
 
-        draws = game_no + 1 - wins[Player.A] - wins[Player.B]
-
         prefix = f"{datetime_prefix(config)}{game_prefix(config, game_no)}"
 
-        print(f"{prefix}Game State:\n{result.board_str}")
-        print(f"{prefix}Game History: {result.history}")
+        if not config.args.concise:
+            print(f"{prefix}Game State:\n{result.board_str}")
+            print(f"{prefix}Game History: {result.history}")
+
         print(f"{prefix}Game Finished: a={result.color[Player.A]}, b={result.color[Player.B]}, win={result.winner}, "
               f"abd={wins[Player.A]}-{wins[Player.B]}-{draws}, bwd={wins[Color.BLACK]}-{wins[Color.WHITE]}-{draws}")
         print(f"{prefix}ELO Updated: a{elo_delta:+}, b{-elo_delta:+}, a={elo[Player.A]}, b={elo[Player.B]}")
@@ -290,6 +324,7 @@ def main():
         f"abd={wins[Player.A]}-{wins[Player.B]}-{draws}, "
         f"awr={wins[Player.A] / config.args.num_games * 100.0}%, bwr={wins[Player.B] / config.args.num_games * 100.0}%, "
         f"bwd={wins[Color.BLACK]}-{wins[Color.WHITE]}-{draws}")
+
 
 if __name__ == "__main__":
     main()
